+++ src/libTAS/frame.h
@@ -36,11 +36,14 @@ extern unsigned long framecount;
  * It is mainly called during a screen refresh (drawFB == true),
  * but can be called also by the timer when we need to advance time to avoid a
  * game softlock (game expect time to pass).
+ * Parameter restore_screen tells if we must redraw the original screen at the
+ * end of the frame boundary. Useful for draw commands that only update part
+ * of the screen.
  */
 #ifdef LIBTAS_ENABLE_HUD
-void frameBoundary(bool drawFB, std::function<void()> draw, RenderHUD& hud);
+void frameBoundary(bool drawFB, std::function<void()> draw, RenderHUD& hud, bool restore_screen);
 #else
-void frameBoundary(bool drawFB, std::function<void()> draw);
+void frameBoundary(bool drawFB, std::function<void()> draw, bool restore_screen);
 #endif
 
 }