+++ src/libTAS/renderhud/RenderHUD_SDL1.cpp
@@ -31,8 +31,11 @@ namespace orig {
     static SDL1::SDL_Surface *(*SDL_CreateRGBSurfaceFrom)(void *pixels,
             int width, int height, int depth, int pitch,
             Uint32 Rmask, Uint32 Gmask, Uint32 Bmask, Uint32 Amask);
+    static void (*SDL_FreeSurface)(SDL1::SDL_Surface *surface);
     static SDL1::SDL_Surface *(*SDL_GetVideoSurface)(void);
     static int (*SDL_BlitSurface)(SDL1::SDL_Surface *src, SDL1::SDL_Rect *srcrect, SDL1::SDL_Surface *dst, SDL1::SDL_Rect *dstrect);
+    static uint8_t (*SDL_SetClipRect)(SDL1::SDL_Surface *surface, const SDL1::SDL_Rect *rect);
+    static void (*SDL_GetClipRect)(SDL1::SDL_Surface *surface, SDL1::SDL_Rect *rect);
 }
 
 RenderHUD_SDL1::~RenderHUD_SDL1()
@@ -42,8 +45,11 @@ RenderHUD_SDL1::~RenderHUD_SDL1()
 void RenderHUD_SDL1::renderText(const char* text, Color fg_color, Color bg_color, int x, int y)
 {
     LINK_NAMESPACE_SDL1(SDL_CreateRGBSurfaceFrom);
+    LINK_NAMESPACE_SDL1(SDL_FreeSurface);
     LINK_NAMESPACE_SDL1(SDL_GetVideoSurface);
     LINK_NAMESPACE_SDL1(SDL_BlitSurface);
+    LINK_NAMESPACE_SDL1(SDL_SetClipRect);
+    LINK_NAMESPACE_SDL1(SDL_GetClipRect);
 
     std::unique_ptr<SurfaceARGB> surf = createTextSurface(text, fg_color, bg_color);
     SDL1::SDL_Surface* sdlsurf = orig::SDL_CreateRGBSurfaceFrom(surf->pixels.data(), surf->w, surf->h, 32, surf->pitch, 0x00FF0000, 0x0000FF00, 0x000000FF, 0xFF000000);
@@ -55,7 +61,15 @@ void RenderHUD_SDL1::renderText(const char* text, Color fg_color, Color bg_color
     y = (y + surf->h + 5) > screen->h ? (screen->h - surf->h - 5) : y;
 
     SDL1::SDL_Rect rect = {static_cast<Sint16>(x), static_cast<Sint16>(y), 0, 0}; // width and height are ignored
+
+    /* Save and restore the clip rectangle */
+    SDL1::SDL_Rect clip_rect;
+    orig::SDL_GetClipRect(screen, &clip_rect);
+    orig::SDL_SetClipRect(screen, nullptr);
     orig::SDL_BlitSurface(sdlsurf, NULL, screen, &rect);
+    orig::SDL_SetClipRect(screen, &clip_rect);
+
+    orig::SDL_FreeSurface(sdlsurf);
 }
 
 }